app/requirements.txt
streamlit>=1.32.0
Pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
opencv-python-headless>=4.5.0
matplotlib>=3.5.0
